+++ lib/rails_admin/config/actions/bulk_delete.rb
@@ -20,7 +20,7 @@ module RailsAdmin
               @objects = list_entries(@model_config, :destroy)
 
               if @objects.blank?
-                flash[:error] = t('admin.flash.error', name: pluralize(0, @model_config.label), action: t('admin.actions.delete.done'))
+                flash[:error] = t('admin.flash.error', name: pluralize(0, @model_config.label, @model_config.label_plural), action: t('admin.actions.delete.done'))
                 redirect_to index_path
               else
                 render @action.template_name
@@ -44,10 +44,10 @@ module RailsAdmin
               end
 
               if destroyed.nil?
-                flash[:error] = t('admin.flash.error', name: pluralize(0, @model_config.label), action: t('admin.actions.delete.done'))
+                flash[:error] = t('admin.flash.error', name: pluralize(0, @model_config.label, @model_config.label_plural), action: t('admin.actions.delete.done'))
               else
-                flash[:success] = t('admin.flash.successful', name: pluralize(destroyed.count, @model_config.label), action: t('admin.actions.delete.done')) unless destroyed.empty?
-                flash[:error] = t('admin.flash.error', name: pluralize(not_destroyed.count, @model_config.label), action: t('admin.actions.delete.done')) unless not_destroyed.empty?
+                flash[:success] = t('admin.flash.successful', name: pluralize(destroyed.count, @model_config.label, @model_config.label_plural), action: t('admin.actions.delete.done')) unless destroyed.empty?
+                flash[:error] = t('admin.flash.error', name: pluralize(not_destroyed.count, @model_config.label, @model_config.label_plural), action: t('admin.actions.delete.done')) unless not_destroyed.empty?
               end
               redirect_to back_or_index
             end